Enable backend to publish messages to MQTT bus
